Absorb Media component into post-fields
